Image size in a uint16_t. 65535x65535 will simply never happen, and that allows casting to a int32_t without running the risk of an integer overflow.

+++ src/gfx.h
@@ -52,8 +52,8 @@ typedef struct ss{
 typedef struct image Image;
 struct image
 {
-  uint32_t height;      // Hauteur
-  uint32_t width;       // Largeur
+  uint16_t height;      // Hauteur
+  uint16_t width;       // Largeur
   SDL_Surface* surface; // L'image en elle mème
 };
